test(plugin-fastify): improvements and split files
